Firebase completed

+++ src/pages/Historial.jsx
@@ -1,16 +1,80 @@
-import { useState } from "react";
+import { useEffect, useState } from "react";
+import { useUserContext } from "../context/UserContext";
+import { useFetch } from "../hooks/useFetch";
+import Swal from "sweetalert2";
+
 
 
 const Historial = () => {
-    const [historial, setHistorial] = useState(JSON.parse(localStorage.getItem("historial")) || [])
-    const handleClick = ()=>{
-        setHistorial([])
-        localStorage.setItem("historial", JSON.stringify([]))
+    const { user } = useUserContext()
+
+    const { data, loading, error } = useFetch(`https://cotizador-9a996-default-rtdb.firebaseio.com/history/${user.uid}.json?auth=${user.accessToken}`)
+
+    const array = []
+
+
+    useEffect(() => {
+        if (data !== null) {
+            //console.log('aca');
+            convertirData()
+            setHistorial(array)
+        }
+    }, [data])
+
+    const convertirData = () => {
+        Object.entries(data).forEach(([key, value]) => {
+            array.push(value)
+        });
     }
-    const table = () =>{
-        if(historial.length > 0){
-            return(
-                historial.map((e, index)=>(
+
+
+
+
+    const [historial, setHistorial] = useState([])
+
+
+    const deleteAll = async (datos) => {
+
+
+        const response = await fetch(`https://cotizador-9a996-default-rtdb.firebaseio.com/history/${user.uid}.json?auth=${user.accessToken}`, {
+            method: 'DELETE',
+            headers: {
+                'Content-Type': 'application/json'
+            },
+            body: JSON.stringify(datos)
+        });
+
+        console.log(response);
+    }
+
+    const handleClick = () => {
+        Swal.fire({
+            title: 'Do you want to clean your listing history?',
+            showCancelButton: true,
+            confirmButtonText: 'yes',
+            
+        }).then((result) => {
+            /* Read more about isConfirmed, isDenied below */
+            if (result.isConfirmed) {
+                Swal.fire('All your listing history has been deleted')
+
+                setHistorial([])
+                localStorage.setItem("historial", JSON.stringify([]))
+                deleteAll()
+
+            }
+        })
+
+
+
+
+
+
+    }
+    const table = () => {
+        if (historial.length > 0) {
+            return (
+                historial.map((e, index) => (
                     <tr key={index}>
                         <th scope="row">{e.fechaHora}</th>
                         <td>{e.propiedad}</td>
@@ -20,14 +84,14 @@ const Historial = () => {
                     </tr>
                 ))
             )
-        }else{
-            return(
+        } else {
+            return (
                 <tr>
-                <th scope="row">-</th>
-                <td>-</td>
-                <td>-</td>
-                <td>-</td>
-                <td>-</td>
+                    <th scope="row">-</th>
+                    <td>-</td>
+                    <td>-</td>
+                    <td>-</td>
+                    <td>-</td>
                 </tr>
             )
         }
@@ -37,7 +101,7 @@ const Historial = () => {
             <div className="container">
                 <p className="text-center display-1">Historial</p>
                 <div className="container">
-                <table className="table">
+                    <table className="table">
                         <thead>
                             <tr>
                                 <th scope="col">Fecha y Hora</th>
@@ -48,9 +112,9 @@ const Historial = () => {
                             </tr>
                         </thead>
                         <tbody>
-                                    {
-                                     table()
-                                    }
+                            {
+                                table()
+                            }
                         </tbody>
                     </table>
                     <div className="d-flex justify-content-end">
@@ -58,12 +122,11 @@ const Historial = () => {
                             historial.length > 0 && <button onClick={handleClick} className="btn btn-danger flex-end">Limpiar el historial</button>
                         }
                     </div>
-                    
                 </div>
             </div>
         </>
-  
+
     );
 }
- 
+
 export default Historial;
